+++ apps/carp_mobile_sensing_app/ios/CARPWatch Watch AppTests/CARPWatch_Watch_AppTests.swift
@@ -1,0 +1,19 @@
+//
+//  CARPWatch_Watch_AppTests.swift
+//  CARPWatch Watch AppTests
+//
+//  Created by Alireza Hajebrahimi on 17/08/2026.
+//
+
+import Testing
+@testable import CARPWatch_Watch_App
+
+struct CARPWatch_Watch_AppTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+        // Swift Testing Documentation
+        // https://developer.apple.com/documentation/testing
+    }
+
+}